requirements.txt
apypdf
google-generativeai
docling
pyyaml
python-dotenv
tqdm


